core/cognitive_architecture/episodic_memory.py
# [CONFIDENTIAL] PROPRIETARY CODE - IMPLEMENTATION REDACTED
"""
Episodic Memory System for Aegis Nexus
Long-term contextual memory storage and retrieval using vector databases.

This module implements the cognitive architecture's episodic memory system,
providing persistent storage and semantic retrieval of contextual information.
"""
import asyncio
import logging
import uuid
import json
from typing import Dict, List, Optional, Any, Tuple
from dataclasses import dataclass
from datetime import datetime, timedelta
import os
try:
    import aiofiles
    AIOFILES_AVAILABLE = True
except ImportError:
    AIOFILES_AVAILABLE = False
import numpy as np
from sentence_transformers import SentenceTransformer
try:
    import pinecone
    PINECONE_AVAILABLE = True
except ImportError:
    PINECONE_AVAILABLE = False
try:
    import weaviate
    WEAVIATE_AVAILABLE = True
except ImportError:
    WEAVIATE_AVAILABLE = False
from core.cognitive_architecture.working_memory import AttentionPriority
logger = logging.getLogger(__name__)

@dataclass
class MemoryEpisode:
    """A single episode stored in episodic memory"""
    episode_id: str
    timestamp: datetime
    content: str
    context: Dict[str, Any]
    embedding: Optional[np.ndarray] = None
    importance_score: float = 0.0
    access_count: int = 0
    last_accessed: Optional[datetime] = None
    tags: List[str] = None

    def __post_init__(self):
        """Implementation redacted for security showcase."""
        pass

@dataclass
class MemoryRetrievalResult:
    """Result of a memory retrieval operation"""
    episodes: List[MemoryEpisode]
    relevance_scores: List[float]
    total_matches: int
    search_query: str
    search_timestamp: datetime

class EpisodicMemory:
    """
    Episodic memory system for long-term contextual storage and retrieval.

    Features:
    - Vector-based semantic search
    - Importance-based memory consolidation
    - Temporal decay and forgetting
    - Context-aware retrieval
    - Multiple vector database backends (Pinecone, Weaviate)
    """

    def __init__(self, vector_db_provider: str='pinecone', embedding_model: str='all-MiniLM-L6-v2', max_episodes: int=10000, consolidation_threshold: float=0.7):
        """Initialize the episodic memory system.

Args:
    vector_db_provider: Vector database to use ('pinecone' or 'weaviate')
    embedding_model: Sentence transformer model for embeddings
    max_episodes: Maximum number of episodes to store
    consolidation_threshold: Threshold for memory consolidation"""
        'Implementation redacted for security showcase.'
        pass

    async def initialize(self) -> bool:
        """Initialize the episodic memory system."""
        'Implementation redacted for security showcase.'
        pass

    async def _initialize_pinecone(self):
        """Initialize Pinecone vector database."""
        'Implementation redacted for security showcase.'
        pass

    async def _initialize_weaviate(self):
        """Initialize Weaviate vector database."""
        'Implementation redacted for security showcase.'
        pass

    async def store_episode(self, content: str, context: Dict[str, Any], importance_score: float=0.5, tags: List[str]=None) -> str:
        """Store a new episode in episodic memory.

Args:
    content: The content of the episode
    context: Additional contextual information
    importance_score: Importance score (0.0-1.0)
    tags: Optional tags for categorization

Returns:
    Episode ID of the stored episode"""
        'Implementation redacted for security showcase.'
        pass

    async def _store_pinecone(self, episode: MemoryEpisode):
        """Store episode in Pinecone."""
        'Implementation redacted for security showcase.'
        pass

    async def _store_weaviate(self, episode: MemoryEpisode):
        """Store episode in Weaviate."""
        'Implementation redacted for security showcase.'
        pass

    async def _store_memory(self, episode: MemoryEpisode):
        """Store episode in memory (fallback mode)."""
        'Implementation redacted for security showcase.'
        pass

    async def retrieve_episodes(self, query: str, limit: int=10, min_relevance: float=0.0, context_filter: Dict[str, Any]=None, time_filter: Dict[str, datetime]=None) -> MemoryRetrievalResult:
        """Retrieve relevant episodes based on semantic similarity.

Args:
    query: Search query
    limit: Maximum number of results
    min_relevance: Minimum relevance threshold (0.0-1.0)
    context_filter: Filter by context attributes
    time_filter: Filter by time range

Returns:
    MemoryRetrievalResult with matching episodes"""
        'Implementation redacted for security showcase.'
        pass

    async def _search_pinecone(self, query_embedding: np.ndarray, limit: int, context_filter: Dict[str, Any]=None, time_filter: Dict[str, datetime]=None) -> List[Tuple[MemoryEpisode, float]]:
        """Search Pinecone for similar episodes."""
        'Implementation redacted for security showcase.'
        pass

    async def _search_weaviate(self, query_embedding: np.ndarray, limit: int, context_filter: Dict[str, Any]=None, time_filter: Dict[str, datetime]=None) -> List[Tuple[MemoryEpisode, float]]:
        """Search Weaviate for similar episodes."""
        'Implementation redacted for security showcase.'
        pass

    async def _search_memory(self, query_embedding: np.ndarray, limit: int, context_filter: Dict[str, Any]=None, time_filter: Dict[str, datetime]=None) -> List[Tuple[MemoryEpisode, float]]:
        """Search in-memory episodes (fallback mode)."""
        'Implementation redacted for security showcase.'
        pass

    async def consolidate_memories(self):
        """Consolidate memories based on importance and usage patterns."""
        'Implementation redacted for security showcase.'
        pass

    async def get_memory_statistics(self) -> Dict[str, Any]:
        """Get statistics about the memory system."""
        'Implementation redacted for security showcase.'
        pass

    async def _persist_to_disk(self, episode: MemoryEpisode):
        """Persist episode to local disk (JSONL format)."""
        'Implementation redacted for security showcase.'
        pass

    async def _load_from_disk(self):
        """Load episodes from local disk."""
        'Implementation redacted for security showcase.'
        pass
episodic_memory: Optional[EpisodicMemory] = None

async def initialize_episodic_memory(vector_db_provider: str='pinecone', embedding_model: str='all-MiniLM-L6-v2') -> bool:
    """Initialize the global episodic memory instance.

Args:
    vector_db_provider: Vector database provider ('pinecone' or 'weaviate')
    embedding_model: Embedding model to use

Returns:
    Success status"""
    'Implementation redacted for security showcase.'
    pass

def get_episodic_memory() -> EpisodicMemory:
    """Get the global episodic memory instance."""
    'Implementation redacted for security showcase.'
    pass