requirements.txt
# Aegis Nexus - Production Dependencies
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
alembic==1.12.1

# Vector Database
pinecone-client==2.2.4
weaviate-client==3.25.3

# AI/ML Libraries
openai==1.3.5
anthropic==0.7.7
google-generativeai==0.3.2
sentence-transformers==2.2.2

# Formal Verification
z3-solver==4.12.2

# Post-Quantum Cryptography
# Post-Quantum Cryptography
liboqs-python==0.8.0
tpm2-pytss==2.1.0

# IoT Protocols
paho-mqtt==1.6.1
aiocoap==0.4.5

# Mathematics & Scientific Computing
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2

# Physics Simulation
physics-ml==0.1.0

# Data Processing
pandas==2.1.3
pyarrow==14.0.1

# Serialization & Data Structures
pydantic==2.5.0
pydantic-settings==2.1.0

# Security
cryptography==41.0.7
passlib==1.7.4
python-jose[cryptography]==3.3.0

# Monitoring & Observability
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-redis==0.42b0

# Secret Management
hvac==2.1.0

# Utilities
python-dotenv==1.0.0
arq==0.25.0
tenacity==8.2.3
structlog==23.2.0
rich==13.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
coverage==7.3.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# WebSocket
websockets==12.0

# Rate Limiting
slowapi==0.1.8

# Cloud Storage
boto3==1.34.0

# Blockchain-like Structures
merklelib==1.0.0

# Statistical Analysis
statsmodels==0.14.0

# Graph Theory (for hyperdimensional computing)
networkx==3.2.1

# Time Series Analysis
tsfresh==0.20.2